renamed Mealplan and ColorScheme
